Modularise Dashboard into separate files.
Modularize Streamlit dashboard into separate files: app.py, data_loader.py, map_utils.py

This is done to allow multiple team members to work on different parts of the dashboard (data loading, map visualization, UI logic) simultaneously without conflicts.
